projects/chapter-23-project/requirements.txt
# Chapter 23 Project Requirements

# Core AI library
openai>=1.0.0

# Logging
structlog>=23.0.0

# Metrics (optional)
# prometheus-client>=0.19.0

# Environment variable management
python-dotenv>=1.0.0
